Implement mechanism to create a cluster instance from an heartbeat

+++ model/cluster.go
@@ -10,7 +10,7 @@ type Scalable interface {
 // ClusterBase is the base class for any type of cluster
 type ClusterBase struct {
 	Name string
-	Nodes int16
+	Nodes int32
 	status Metrics // not available outside package to prevent race conditions- get and set must be used
 	sync.Mutex
 }
@@ -28,7 +28,7 @@ type ClusterBaseInterface interface {
 // @param clusterName is the name of the cluster
 // @param size is the number of nodes in the cluster
 // return the pointer to the ClusterBase instance
-func NewClusterBase(clusterName string, size int16) *ClusterBase {
+func NewClusterBase(clusterName string, size int32) *ClusterBase {
 	return &ClusterBase{
 		Name:  clusterName,
 		Nodes: size,
